+++ src/watchdog/tests/observers/emitter_mixin.py
@@ -1,6 +1,8 @@
 """
 Common tests for all emitters.
 """
+import unicodedata
+
 from watchdog.events import EVENT_TYPE_CREATED, EVENT_TYPE_MODIFIED
 from watchdog.tests import mk
 
@@ -16,31 +18,25 @@ class EmitterSystemMixin(object):
         """
         pass
 
-    def test_run_ok(self):
+    def test_start_ok(self):
         """
-        After emitter start, ready event is set and no error is set.
+        No errors are raised when emitter is successfully started.
         """
         self.sut.start()
 
-        self.assertIsNone(self.sut.start_error)
-        self.assertIsTrue(self.sut.ready.is_set())
         self.assertTrue(self.emitter_queue.empty())
 
         self.endThread(self.sut)
 
-    def test_run_bad_path(self):
+    def test_start_bad_path(self):
         """
         It can be initialized with a bad path but error is only
         raised when emitter starts.
         """
         sut = self.makeEmitter(path='no-such-path')
-        self.assertIsNone(sut.start_error)
 
-        # Start will call the run method.
-        sut.start()
-
-        self.assertIsInstance(OSError, sut.start_error)
-        self.assertIsTrue(sut.ready.is_set())
+        with self.assertRaises(OSError):
+            sut.start()
 
     def test_queue_events_file_created(self):
         """
@@ -58,10 +54,27 @@ class EmitterSystemMixin(object):
         event, watch = self.emitter_queue.get()
         new_path = mk.fs.getEncodedPath(
             mk.fs.getRealPathFromSegments(self.test_segments))
-        self.assertEqual(new_path, event.src_path)
+
+        if self.os_name == 'osx':
+            src_path = event.src_path.decode('utf-8')
+            self.assertEqual(
+                mk.fs.getRealPathFromSegments(self.test_segments),
+                unicodedata.normalize('NFC', src_path),
+                )
+        else:
+            self.assertEqual(new_path, event.src_path)
+
         self.assertFalse(event.is_directory)
         self.assertEqual(EVENT_TYPE_CREATED, event.event_type)
         event, watch = self.emitter_queue.get()
-        self.assertEqual(mk.fs.getEncodedPath(mk.fs.temp_path), event.src_path)
+
+        if self.os_name == 'osx':
+            src_path = event.src_path.decode('utf-8')
+            self.assertEqual(
+                mk.fs.temp_path, unicodedata.normalize('NFC', src_path))
+        else:
+            self.assertEqual(
+                mk.fs.getEncodedPath(mk.fs.temp_path), event.src_path)
+
         self.assertTrue(event.is_directory)
         self.assertEqual(EVENT_TYPE_MODIFIED, event.event_type)