+++ src/watchdog/tests/observers/test_api.py
@@ -19,7 +19,6 @@
 Tests for generic observer, generic emitter, generic event dispatcher...etc
 """
 import tempfile
-import threading
 
 from watchdog.observers.api import BaseObserver, EventEmitter, ObservedWatch
 from watchdog.tests import WatchdogTestCase
@@ -31,22 +30,6 @@ class DummyEventEmitter(EventEmitter):
     event emitters.
     """
 
-    def start(self):
-        self.ready.set()
-        return super(DummyEventEmitter, self).start()
-
-
-class DelayStartEventEmitter(EventEmitter):
-    """
-    An event emitter which delays the start until `delay` member is set.
-    """
-    exception = AssertionError('fail-to-start')
-
-    def start(self):
-        self.delay = threading.Event()
-        self.delay.wait(timeout=10)
-        self.ready.set()
-
 
 class FailToStartEventEmitter(EventEmitter):
     """
@@ -55,10 +38,9 @@ class FailToStartEventEmitter(EventEmitter):
     exception = AssertionError('fail-to-start')
     started = False
 
-    def start(self):
+    def on_thread_start(self):
         self.started = True
-        self._start_error = self.exception
-        self.ready.set()
+        raise self.exception
 
 
 class TestEventEmitter(WatchdogTestCase):
@@ -84,7 +66,7 @@ class TestEventEmitter(WatchdogTestCase):
         """
         self.sut.start()
 
-        self.assertTrue(self.sut.ready.is_set())
+        self.assertTrue(self.sut.is_alive())
 
 
 class DummyObserver(BaseObserver):
@@ -116,7 +98,7 @@ class TestBaseObserver(WatchdogTestCase):
         if observer is None:
             observer = self.sut
         try:
-            return observer._get_emitter_for_watch(watch)
+            return observer._emitter_for_watch.get(watch)
         except KeyError:
             return None
 
@@ -140,7 +122,6 @@ class TestBaseObserver(WatchdogTestCase):
 
         emitter = self.getEmitterForWatch(handler)
         self.assertTrue(emitter.is_alive())
-        self.assertTrue(emitter.ready)
 
     def test_schedule_started_fail(self):
         """
@@ -168,11 +149,10 @@ class TestBaseObserver(WatchdogTestCase):
         self.assertEqual(2, len(self.sut._emitters))
 
         with self.assertRaises(AssertionError) as context:
-            self.sut.run()
+            self.sut.start()
 
         self.assertIs(FailToStartEventEmitter.exception, context.exception)
         self.assertEqual(1, len(self.sut._emitters))
-        self.assertTrue(self.sut.ready.is_set())
 
     def test_start_blocking(self):
         """
@@ -183,28 +163,10 @@ class TestBaseObserver(WatchdogTestCase):
         self.assertFalse(emitter.is_alive())
         self.assertFalse(self.sut.is_alive())
 
-        self.sut.start_blocking()
+        self.sut.start()
 
         self.assertTrue(emitter.is_alive())
-        self.assertTrue(emitter.ready.is_set())
         self.assertTrue(self.sut.is_alive())
-        self.assertTrue(self.sut.ready.is_set())
-
-    def test_start_blocking_timeout(self):
-        """
-        It starts the attached emitters.
-        """
-        self.sut._emitter_class = DelayStartEventEmitter
-        handler = self.sut.schedule('handler', tempfile.tempdir)
-        emitter = self.getEmitterForWatch(handler)
-        self.assertFalse(self.sut.is_alive())
-
-        with self.assertRaises(AssertionError) as context:
-            self.sut.start_blocking(timeout=0.01)
-
-        emitter.delay.set()
-        self.assertEqual('Took to much to start.', context.exception.message)
-        self.assertFalse(self.sut.ready.is_set())
 
     def test_unschedule_not_started(self):
         """