+++ src/watchdog/utils/__init__.py
@@ -24,7 +24,7 @@
 
 Classes
 -------
-.. autoclass:: DaemonThread
+.. autoclass:: BaseThread
    :members:
    :show-inheritance:
    :inherited-members:
@@ -34,6 +34,7 @@ import os
 import sys
 import threading
 import watchdog.utils.platform
+from watchdog.utils.compat import Event
 from collections import namedtuple
 
 
@@ -72,12 +73,8 @@ def has_attribute(ob, attribute):
     return getattr(ob, attribute, None) is not None
 
 
-class DaemonThread(threading.Thread):
-
-    """
-    Daemon thread convenience class, sets a few properties and makes
-    writing daemon threads a little easier.
-    """
+class BaseThread(threading.Thread):
+    """ Convenience class for creating stoppable threads. """
 
     def __init__(self):
         threading.Thread.__init__(self)
@@ -85,7 +82,7 @@ class DaemonThread(threading.Thread):
             self.daemon = True
         else:
             self.setDaemon(True)
-        self._stopped_event = threading.Event()
+        self._stopped_event = Event()
 
         if not has_attribute(self._stopped_event, 'is_set'):
             self._stopped_event.is_set = self._stopped_event.isSet
@@ -95,22 +92,34 @@ class DaemonThread(threading.Thread):
         return self._stopped_event
 
     def should_keep_running(self):
-        """Determines whether the daemon thread should continue running."""
+        """Determines whether the thread should continue running."""
         return not self._stopped_event.is_set()
 
     def on_thread_stop(self):
         """Override this method instead of :meth:`stop()`.
         :meth:`stop()` calls this method.
 
-        Note that this method is called immediately after the daemon thread
-        is signaled to halt.
+        This method is called immediately after the thread is signaled to stop.
         """
         pass
 
     def stop(self):
-        """Signals the daemon thread to stop."""
+        """Signals the thread to stop."""
         self._stopped_event.set()
         self.on_thread_stop()
+
+    def on_thread_start(self):
+        """Override this method instead of :meth:`start()`. :meth:`start()`
+        calls this method.
+
+        This method is called right before this thread is started and this
+        object’s run() method is invoked.
+        """
+        pass
+
+    def start(self):
+        self.on_thread_start()
+        threading.Thread.start(self)
 
 
 def load_module(module_name):