+++ tests/test_emitter.py
@@ -43,19 +43,10 @@ def setup_function(function):
     event_queue = Queue()
 
 
-def make_emitter(path=None):
-    """
-    Create the default emitter, return it and set it as global to be
-    available in teardown_function.
-    """
+def start_watching(path=None):
     path = p('') if path is None else path
     global emitter
     emitter = Emitter(event_queue, ObservedWatch(path, recursive=True))
-    return emitter
-
-
-def start_watching(path=None):
-    emitter = make_emitter(path)
     if platform.is_darwin():
         # FSEvents will report old evens (like create for mkdtemp in test
         # setup. Waiting for a considerable time seems to 'flush' the events.
@@ -65,8 +56,9 @@ def start_watching(path=None):
 
 def teardown_function(function):
     emitter.stop()
-    emitter.join()
+    emitter.join(5)
     rm(p(''), recursive=True)
+    assert not emitter.is_alive()
 
 
 def test_create():
@@ -195,21 +187,3 @@ def test_passing_bytes_should_give_bytes():
     touch(p('a'))
     event = event_queue.get(timeout=5)[0]
     assert isinstance(event.src_path, bytes)
-
-
-def test_ignore_events_before_start():
-    """
-    It will ignore events occurring on the FS before start is called.
-    """
-    emitter = make_emitter(p(''))
-    touch(p('a'))
-
-    assert event_queue.empty()
-
-    # Now we start the emitter and after this we should see events.
-    emitter.start()
-    touch(p('b'))
-
-    event = event_queue.get(timeout=5)[0]
-    assert isinstance(event, FileCreatedEvent)
-    assert event.src_path == p('b')